Moved validation functions to `validation.js`
