+++ static/js/components/CourseList_test.js
@@ -8,7 +8,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 
 import CourseList from './CourseList';
-import * as util from '../util/util';
+import * as courseList from '../util/courseList';
 import { STATUS_PASSED, DASHBOARD_RESPONSE } from '../constants';
 
 describe("CourseList", () => {
@@ -61,8 +61,8 @@ describe("CourseList", () => {
   });
 
   it('renders courses using util functions', () => {
-    let makeCourseProgressDisplaySpy = sandbox.spy(util, 'makeCourseProgressDisplay');
-    let makeCourseStatusDisplaySpy = sandbox.spy(util, 'makeCourseStatusDisplay');
+    let makeCourseProgressDisplaySpy = sandbox.spy(courseList, 'makeCourseProgressDisplay');
+    let makeCourseStatusDisplaySpy = sandbox.spy(courseList, 'makeCourseStatusDisplay');
 
     TestUtils.renderIntoDocument(
       <MuiThemeProvider muiTheme={getMuiTheme()}>
